pyproject.toml
[project]
name = "data-power"
version = "0.1.0"
description = "Test task for Data Power"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
